making test suite for Chef::Provider::DockerService::Execute pass

+++ libraries/provider_docker_service_execute.rb
@@ -10,6 +10,13 @@ class Chef
         action :start do
           action_stop unless resource_changes.empty?
 
+          # enable ipv6 forwarding
+          execute 'enable net.ipv6.conf.all.forwarding' do
+            command '/sbin/sysctl net.ipv6.conf.all.forwarding=1'
+            not_if '/sbin/sysctl -q -n net.ipv6.conf.all.forwarding | grep ^1$'
+            action :run
+          end
+
           # Go doesn't support detaching processes natively, so we have
           # to manually fork it from the shell with &
           # https://github.com/docker/docker/issues/2758
@@ -22,12 +29,21 @@ class Chef
             not_if "ps -ef | awk '{ print $8 }' | grep ^#{docker_bin}$"
             action :run
           end
+
+          # loop until docker socker is available
+          ruby_block 'wait for docker' do
+            block do
+              true until ::File.exist?('/var/run/docker.sock')
+            end
+            not_if { ::File.exist? '/var/run/docker.sock' }
+          end
         end
 
         action :stop do
           execute 'stop docker' do
             command 'kill `pidof docker`'
             only_if "ps -ef | awk '{ print $8 }' | grep ^#{docker_bin}$"
+            not_if { ::File.exist? '/var/run/docker.sock' }
           end
         end
